+++ ourWinch/Controllers/Dashboard/ActiveServiceController.cs
@@ -13,25 +13,8 @@ public class ActiveServiceController : Controller
     [HttpPost]
     public IActionResult SaveToActiveService(int serviceOrderId)
     {
-        var serviceOrder = _context.ServiceOrders.FirstOrDefault(so => so.ServiceOrderId == serviceOrderId);
-        if (serviceOrder != null)
-        {
-            var activeService = new ActiveService
-            {
-                Ordrenummer = serviceOrder.Ordrenummer,
-                Produkttype = serviceOrder.Produkttype,
-                Fornavn = serviceOrder.Fornavn,
-                Etternavn = serviceOrder.Etternavn,
-                MottattDato = serviceOrder.MottattDato,
-                Feilbeskrivelse = serviceOrder.Feilbeskrivelse,           
-                Status = "Process", // Bu özelliği varsayılan olarak "Process" olarak atadım.   
-            };
-            System.Diagnostics.Debug.WriteLine("Before Save: " + activeService.Ordrenummer);
-            _context.ActiveServices.Add(activeService);
-            _context.SaveChanges();
-            System.Diagnostics.Debug.WriteLine("After Save: " + activeService.Ordrenummer);
-        }
-
+        var serviceManager = new ServiceManager(_context);
+        serviceManager.SaveToActiveService(serviceOrderId);
         return RedirectToAction("Index");
     }
 }
